+++ src/app/page-component/search-bar/search-bar.component.ts
@@ -1,0 +1,18 @@
+import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
+
+@Component({
+  selector: 'app-search-bar',
+  templateUrl: './search-bar.component.html',
+  styleUrls: ['./search-bar.component.css']
+})
+export class SearchBarComponent implements OnInit {
+  @Output() text = new EventEmitter();
+
+  constructor() { }
+  ngOnInit() { }
+
+
+  onKey(event: any) {
+    this.text.emit(event.target.value)
+  }
+}