+++ Changelog.md
@@ -5,6 +5,7 @@
     * Add "note" attr for image: annotation(:string) of image
     * Add copy button to all result_sets
     * Improved  performance of result_set list
+    * Add search button
 ### Fixed
     * Fixed race condition while plans load
 ### Changed